Added code for endpoint /profile, POST, so user can now update their profile.
Currently, it is not working as intended in all forms.
The user has to enter in both password1 and password2 in order to update the profile. Otherwise, it fails.
Also, until I figure out how to extract the password from the authentication, I am not sure how to verify the entered password, is the correct password.
Updating the profile should not allow the user to change their password in this manner without some sort of other security check.